agenda/requirements/local.txt
Django>=3.0.5,<=3.1.0
django-model-utils>=4.0.0,<=4.1.0
psycopg2-binary>=2.9.6,<=3.0.0
Unipath>=1.1,<=1.2
flake8>=6.0.0,<=6.7.0
djangorestframework>=3.14.0,<=3.15.0
Faker>=18.4.0,<=18.5.0
requests>=2.28.2,<=2.29.0
